Add prismjs theme to webpack
And add to app.blade.php the CSS file.
The problem that in Question can't display hightlight syntax is because it is surrounded with clean() (in Model Question.php).
and clean() comes from purifier. then it need to fix a few things in purifier.php.
the highlight syntax display with pre and code tag.